+++ CHANGELOG.md
@@ -70,6 +70,7 @@
 * Add `EveOnline::ESI::AllianceIcon`
 * Add `EveOnline::ESI::Models::Alliance`
 * Add `EveOnline::ESI::Models::AllianceIcon`
+* Add `EveOnline::ESI::Models::CharacterPortrait`
 
 **v0.12.0**